Move SimplifiedRelationType to resulttypes.ts
